rdmotorsAPI/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
PyMySQL==1.1.2
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
python-dotenv==1.0.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-flask==1.3.0
pytest-mock==3.14.0

# Security
Flask-Limiter==3.8.0
bleach==6.1.0
cryptography==43.0.3

# API Documentation
flask-restx==1.3.0

# Production
gunicorn==23.0.0

